requirements.txt
# Core framework dependencies
google-adk>=0.1.0
asyncio
pydantic>=2.0.0
typing-extensions

# MCP protocol dependencies  
mcp>=0.1.0

# Testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.28.0
aiohttp>=3.8.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0